+++ src/Exceptions/RequestDeniedException.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Gowelle\LaravelRouteMatrix\Exceptions;
+
+/**
+ * Exception thrown when the API request is denied (HTTP 403).
+ */
+class RequestDeniedException extends GoogleRoutesException
+{
+    //
+}